requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
absl-py==1.1.0
accelerate==0.11.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
Deprecated==1.2.13
docker-pycreds==0.4.0
einops==0.4.1
einops-exts==0.0.3
ema-pytorch==0.0.8
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==3.7.0
humanize==4.2.2
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
joblib==1.1.0
kiwisolver==1.4.3
lxml==4.9.1
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
networkx==2.8.5
nibabel==4.0.1
nilearn==0.9.1
numpy==1.23.0
oauthlib==3.2.0
omegaconf==2.2.3
pandas==1.4.3
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyDeprecate==0.3.1
pydicom==2.3.0
pytorch-lightning==1.6.4
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==19.0.2
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rotary-embedding-torch==0.1.5
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-video==1.1.11
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
SimpleITK==2.1.1.2
smmap==5.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.8.3
toml==0.10.2
torch-tb-profiler==0.4.0
torchio==0.18.80
torchmetrics==0.9.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.21
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
pytorch-msssim
